requirements.txt
# Unified dependency file for DualComm
# Includes python_bridge runtime dependencies + project RAG/indexing tools.

# Core Python API runtime (python_bridge)
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
orjson>=3.10,<4.0
python-multipart>=0.0.9,<1.0

# LLM / agent runtime integrations
langchain>=0.2,<1.0
langchain-community>=0.2.0
langchain-openai>=0.1,<1.0
langchain-groq>=0.1,<1.0
groq>=0.9,<1.0

# RAG / vector stack
cohere>=5.0.0
qdrant-client>=1.9.0
llama-index>=0.11,<1.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-cohere>=0.1.0

# Knowledge base ingestion and parsing
pymupdf>=1.24.0
requests>=2.31.0
pillow>=10.0.0

# Translation
transformers>=4.40.0
sentencepiece>=0.2.0
torch>=2.2.0

# Optional channel integrations
python-telegram-bot>=21.0

# Shared utility
python-dotenv>=1.0,<2.0

# Advocacy & MCP (Required for Government Advocacy)
resend
fpdf2
mcp
